Added part of the Red Bat enemy and it's banner
Some glitches are now occurring that I can't seem to fix, but I also had
to change some other cs files, due to that some variables where created
but were never used, which made tModLoader refused to build the mod for
me, and now the Machine gun no longer displays on the cursor and moves
around quickly when in use. Also, placing the "Red Bat Banner" just
places a dirt block. I'll be waiting for a patch, I'm new to C#
programming.